third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-overlapping-elements-expected.txt

This test ensures that Spatial Navigation works with overlapping elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Element 1Element 2Element 3Element 4Element 5Element 6
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e4"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e5"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e6"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e5"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e4"
PASS successfullyParsed is true

TEST COMPLETE
